+++ components/spider-execution-manager/src/client.rs
@@ -6,10 +6,12 @@
 //! * [`storage::StorageClient`] — registers task instances and reports their outcome.
 //! * [`liveness::LivenessClient`] — registers the EM at boot and ticks the heartbeat thereafter.
 
+pub mod grpc;
 pub mod liveness;
 pub mod scheduler;
 pub mod storage;
 
+pub use grpc::GrpcStorageClient;
 pub use liveness::{LivenessClient, LivenessResponseError, RegistrationResponse};
 pub use scheduler::{SchedulerClient, SchedulerError, SchedulerResponse};
-pub use storage::{GrpcStorageClient, StorageClient, StorageResponseError};
+pub use storage::{StorageClient, StorageResponseError};
