+++ components/spider-execution-manager/src/client/grpc/mod.rs
@@ -1,0 +1,5 @@
+//! gRPC-backed implementations of the execution manager's client traits.
+
+pub mod storage;
+
+pub use storage::GrpcStorageClient;
