+++ components/spider-execution-manager/src/client/storage.rs
@@ -8,17 +8,6 @@ use spider_core::types::{
     id::{ExecutionManagerId, JobId, SessionId, TaskId},
     io::ExecutionContext,
 };
-use spider_proto_rust::{
-    id::task_id_to_protocol,
-    storage::{
-        self,
-        register_task_instance_response,
-        storage_error,
-        storage_operation_response,
-        task_instance_management_service_client::TaskInstanceManagementServiceClient,
-    },
-};
-use tonic::transport::{Channel, Endpoint};
 
 /// Errors returned by [`StorageClient`] operations.
 ///
@@ -142,230 +131,4 @@ pub trait StorageClient: Send + Sync {
         session_id: SessionId,
         error_message: String,
     ) -> Result<(), StorageResponseError>;
-}
-
-/// gRPC-backed [`StorageClient`] implementation.
-#[derive(Debug, Clone)]
-pub struct GrpcStorageClient {
-    inner: TaskInstanceManagementServiceClient<Channel>,
-}
-
-impl GrpcStorageClient {
-    /// Connects to the storage gRPC endpoint.
-    ///
-    /// # Returns
-    ///
-    /// A new [`GrpcStorageClient`] connected to `endpoint` on success.
-    ///
-    /// # Errors
-    ///
-    /// Returns an error if:
-    ///
-    /// * [`StorageResponseError::Transport`] if tonic cannot create or connect to the endpoint.
-    pub async fn connect(endpoint: Endpoint) -> Result<Self, StorageResponseError> {
-        TaskInstanceManagementServiceClient::connect(endpoint)
-            .await
-            .map(|inner| Self { inner })
-            .map_err(to_transport_error)
-    }
-}
-
-#[async_trait]
-impl StorageClient for GrpcStorageClient {
-    async fn register_task_instance(
-        &self,
-        job_id: JobId,
-        task_id: TaskId,
-        em_id: ExecutionManagerId,
-        session_id: SessionId,
-    ) -> Result<ExecutionContext, StorageResponseError> {
-        let request = storage::RegisterTaskInstanceRequest {
-            job_id: job_id.get(),
-            task_id: Some(task_id_to_protocol(task_id)),
-            execution_manager_id: em_id.get(),
-            session_id,
-        };
-        let response = {
-            let mut client = self.inner.clone();
-            client
-                .register_task_instance(request)
-                .await
-                .map_err(to_transport_error)?
-                .into_inner()
-        };
-
-        match response.result {
-            Some(register_task_instance_response::Result::ExecutionContext(bytes)) => {
-                bincode::deserialize(&bytes).map_err(|error| {
-                    StorageResponseError::Transport(format!(
-                        "failed to decode execution context: {error}"
-                    ))
-                })
-            }
-            Some(register_task_instance_response::Result::Error(error)) => Err(error.into()),
-            None => Err(StorageResponseError::Transport(
-                "register task instance response missing result".to_owned(),
-            )),
-        }
-    }
-
-    async fn report_task_success(
-        &self,
-        job_id: JobId,
-        task_id: TaskId,
-        em_id: ExecutionManagerId,
-        session_id: SessionId,
-        serialized_outputs: Option<Vec<u8>>,
-    ) -> Result<(), StorageResponseError> {
-        let request = storage::ReportTaskSuccessRequest {
-            job_id: job_id.get(),
-            task_id: Some(task_id_to_protocol(task_id)),
-            execution_manager_id: em_id.get(),
-            session_id,
-            serialized_outputs: serialized_outputs.unwrap_or_default(),
-        };
-        let response = {
-            let mut client = self.inner.clone();
-            client
-                .report_task_success(request)
-                .await
-                .map_err(to_transport_error)?
-                .into_inner()
-        };
-
-        storage_operation_response_to_result(
-            response,
-            "report task success response missing result",
-        )
-    }
-
-    async fn report_task_failure(
-        &self,
-        job_id: JobId,
-        task_id: TaskId,
-        em_id: ExecutionManagerId,
-        session_id: SessionId,
-        error_message: String,
-    ) -> Result<(), StorageResponseError> {
-        let request = storage::ReportTaskFailureRequest {
-            job_id: job_id.get(),
-            task_id: Some(task_id_to_protocol(task_id)),
-            execution_manager_id: em_id.get(),
-            session_id,
-            error_message,
-        };
-        let response = {
-            let mut client = self.inner.clone();
-            client
-                .report_task_failure(request)
-                .await
-                .map_err(to_transport_error)?
-                .into_inner()
-        };
-
-        storage_operation_response_to_result(
-            response,
-            "report task failure response missing result",
-        )
-    }
-}
-
-impl From<storage::StorageError> for StorageResponseError {
-    fn from(error: storage::StorageError) -> Self {
-        match storage_error::ErrCode::try_from(error.err_code) {
-            Ok(storage_error::ErrCode::StaleSession) => Self::StaleSession {
-                storage_session: error.storage_session,
-            },
-            Ok(storage_error::ErrCode::CacheStale) => Self::CacheStale(error.message),
-            Ok(storage_error::ErrCode::Transport) => Self::Transport(error.message),
-            Ok(storage_error::ErrCode::Server | storage_error::ErrCode::Unspecified) => {
-                Self::Server(error.message)
-            }
-            Ok(storage_error::ErrCode::InvalidInput) => Self::InvalidInput(error.message),
-            Err(error) => Self::Transport(format!("unknown storage error kind: {error}")),
-        }
-    }
-}
-
-/// Converts a protobuf storage operation response into the local empty success/error result.
-///
-/// # Returns
-///
-/// Converted gRPC response result if the `response` has a valid result..
-///
-/// # Errors
-///
-/// Returns an error if:
-///
-/// * [`StorageResponseError::Transport`] if the response does not contain a result.
-fn storage_operation_response_to_result(
-    response: storage::StorageOperationResponse,
-    missing_result_message: &'static str,
-) -> Result<(), StorageResponseError> {
-    match response.result {
-        Some(storage_operation_response::Result::Ok(_)) => Ok(()),
-        Some(storage_operation_response::Result::Error(error)) => Err(error.into()),
-        None => Err(StorageResponseError::Transport(
-            missing_result_message.to_owned(),
-        )),
-    }
-}
-
-/// Converts a displayable transport-layer error into [`StorageResponseError::Transport`].
-///
-/// # Returns
-///
-/// A [`StorageResponseError::Transport`] containing `error`'s display string.
-fn to_transport_error(error: impl std::fmt::Display) -> StorageResponseError {
-    StorageResponseError::Transport(error.to_string())
-}
-
-#[cfg(test)]
-mod tests {
-    use super::*;
-
-    #[test]
-    fn storage_error_maps_stale_session() {
-        let error = storage::StorageError {
-            err_code: storage_error::ErrCode::StaleSession.into(),
-            message: "stale".to_owned(),
-            storage_session: 7,
-        };
-
-        match StorageResponseError::from(error) {
-            StorageResponseError::StaleSession { storage_session } => {
-                assert_eq!(7, storage_session);
-            }
-            error => panic!("unexpected storage response error: {error:?}"),
-        }
-    }
-
-    #[test]
-    fn storage_error_maps_unknown_kind_to_transport_error() {
-        let error = storage::StorageError {
-            err_code: 99,
-            message: "unknown".to_owned(),
-            storage_session: 0,
-        };
-
-        match StorageResponseError::from(error) {
-            StorageResponseError::Transport(message) => {
-                assert!(message.contains("unknown storage error kind"));
-            }
-            error => panic!("unexpected storage response error: {error:?}"),
-        }
-    }
-
-    #[test]
-    fn missing_storage_operation_result_is_transport_error() {
-        match storage_operation_response_to_result(
-            storage::StorageOperationResponse { result: None },
-            "missing result",
-        ) {
-            Err(StorageResponseError::Transport(message)) => {
-                assert_eq!("missing result", message);
-            }
-            result => panic!("unexpected storage operation result: {result:?}"),
-        }
-    }
 }
